% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrectDataBasedAmon.R
\name{CorrectDataBasedAmon}
\alias{CorrectDataBasedAmon}
\title{CorrectDataBasedAmon}
\usage{
CorrectDataBasedAmon(inZoo = inZoo, inAmon = inAmon)
}
\arguments{
\item{inZoo}{input time series.}

\item{inAmon}{input anomoly.}
}
\description{
This function calculate daily amonoly from a zoo object.
}
\examples{
CorrectDataBasedAmon()
}
\keyword{data}
