% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DailyAmon.R
\name{DailyAmon}
\alias{DailyAmon}
\title{calculate daily amonoly from a zoo object}
\usage{
DailyAmon(inZoo = inZoo)
}
\arguments{
\item{inZoo}{file to read.}
}
\description{
This function calculate daily amonoly from a zoo object.
}
\examples{
DailyAmon()
}
\keyword{data}
