% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearestXY.R
\name{nearestXY}
\alias{nearestXY}
\title{find nearestXY}
\usage{
nearestXY(x, y, X = X, Y = Y)
}
\arguments{
\item{x}{x coordinate of point lon or utm_e}

\item{y}{y coordinate of point lon or utm_n}

\item{X}{x coordinate of points lon or utm_e}

\item{Y}{y coordinate of points lon or utm_n}
}
\description{
find nearestXY
}
\examples{
nearestXY()
}
\keyword{data}
