% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readHBV.R
\name{readHBV}
\alias{readHBV}
\title{read nve data}
\usage{
readHBV(fileName, na = "NULL")
}
\arguments{
\item{fileName}{file to read.}

\item{na.rm}{if remove the null values, "NULL" not include, "NA" eller "any"; default is "NULL"}
}
\description{
This function read the HBV format data and return a zoo object.
}
\examples{
readHBV(fileName = "data/dew_00400000.out", na = "NULL")
}
\keyword{data}
