% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seNorgeDailyTimeSeriesDay.R
\name{seNorgeDailyTimeSeriesDay}
\alias{seNorgeDailyTimeSeriesDay}
\title{get SeNorge time series for a point (x, y)}
\usage{
seNorgeDailyTimeSeriesDay(x, y, s = s, e = e, var = "tm",
  path = "//hdata/grid/metdata/met_obs_v2.0")
}
\arguments{
\item{x}{x direction coordinate lon, utm_e}

\item{y}{y direction coordinate lat, utm_n}

\item{s}{start date}

\item{e}{end date}

\item{var}{variable to extract: tm or rr}

\item{path}{path to data}
}
\description{
get SeNorge time series for a point (x, y)
}
\examples{
seNorgeDailyTimeSeriesDay()
}
\keyword{data}
