% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seNorgeDailyTimeSeriesYear.R
\name{seNorgeDailyTimeSeriesYear}
\alias{seNorgeDailyTimeSeriesYear}
\title{get SeNorge time series for a point (x, y)}
\usage{
seNorgeDailyTimeSeriesYear(x, y, s = s, e = e, var = "tm",
  path = "//hdata/grid/metdata/met_obs_v2.1")
}
\arguments{
\item{x}{x direction coordinate lon, utm_e}

\item{y}{y direction coordinate lat, utm_n}

\item{s}{start year}

\item{e}{end year}

\item{var}{variable to extract: tm or rr}

\item{path}{path to data}
}
\description{
get SeNorge time series for a point (x, y)
}
\examples{
seNorgeDailyTimeSeriesYear()
}
\keyword{data}
