% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeHBV_disc.R
\name{writeHBV_disc}
\alias{writeHBV_disc}
\title{writeHBV_disc.R}
\usage{
writeHBV_disc(DataZoo, fileName = "data/discharge_data.txt")
}
\arguments{
\item{fileName}{file to write.}
}
\description{
This function write the HBV format discharge data.
}
\examples{
writeHBV_disc()
}
\keyword{data}
