% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeHBV_metinput.R
\name{writeHBV_metinput}
\alias{writeHBV_metinput}
\title{writeHBV_metinput.R}
\usage{
writeHBV_metinput(DataZoo, fileName = "data/input_data.txt")
}
\description{
This function write the HBV format input met data.
}
\examples{
writeHBV_metinput()
}
\keyword{data}
