% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readHBV.R
\name{readHBV}
\alias{readHBV}
\title{read nve data}
\usage{
readHBV(fileName = "data/dew_00400000.out")
}
\arguments{
\item{fileName}{file to read.}
}
\description{
This function read the HBV format data.
}
\examples{
readHBV()
}
\keyword{data}
